requirements.txt
Django==2.2.19
pytz==2022.7.1
sqlparse==0.4.3
python-telegram-bot==13.7